Move PaymentMethodSelector to own file with dependencies
