+++ client/me/purchases/components/payment-method-form/style.scss
@@ -2,8 +2,7 @@
 	margin-bottom: 0;
 }
 
-.payment-method-form__terms,
-.change-payment-method__terms {
+.payment-method-form__terms {
 	color: var( --color-text-subtle );
 	margin: 5px 0 0;
 	padding: 0;
@@ -11,7 +10,6 @@
 	p {
 		font-size: $font-body-extra-small;
 		margin-left: 24px;
-		
 	}
 
 	.gridicon {
@@ -43,38 +41,4 @@
 .payment-method-form__heading {
 	font-size: $font-title-small;
 	padding: 0 0 16px;
-}
-
-/* Override checkout styles */
-.change-payment-method__content {
-	.credit-card-fields-inner-wrapper {
-		max-width: 525px;
-	}
-
-	.contact-fields {
-		margin-top: 16px;
-		padding-bottom: 8px;
-	}
-
-	.checkout__step-wrapper.checkout__step-wrapper--last-step {
-		margin-bottom: 0;
-		border: 0;
-		width: 100%;
-		max-width: none;
-	}
-
-	.checkout__step-wrapper--last-step .checkout-steps__submit-button-wrapper {
-		background: transparent;
-		padding: 0;
-		position: relative;
-		border: 0;
-	}
-
-	.checkout-submit-button .checkout-button {
-		width: auto;
-
-		&:focus {
-			outline: var( --color-primary-light ) solid 2px;
-		}
-	}
 }
