+++ client/me/purchases/manage-purchase/payment-method-selector/tos-text.tsx
